Use Event#sort_column for finding previous events
Now the previous events are always returned in predictable order, and it
is also possible to find the single previous event without instantiating
all previous events.

By the way, the old `this_and_previous_events[-2]` would not always
return the correct event. If multiple events are executed on the same
date, all those events would be part of `#this_and_previous_events`. If
e.g. 3 events were executed on that date, they would be on indexes -3,
-2 and -1. So it's easy to imagine `this_and_previous_events[-2]`
pointing not to the previous event, but in fact to "this" event. It
feels nice to finally have this bug fixed.